requirements.txt
openai>=1.0.0
anthropic>=0.25.0
pyyaml>=6.0
networkx>=3.0
pyverilog>=1.3.0
streamlit>=1.35.0
groq>=0.11.0
python-dotenv>=1.0.1
